+++ addon/components/variable-plugin/number/insert.ts
@@ -6,6 +6,7 @@ import { v4 as uuidv4 } from 'uuid';
 import { isNumber } from '@lblod/ember-rdfa-editor-lblod-plugins/utils/strings';
 import { inject as service } from '@ember/service';
 import IntlService from 'ember-intl/services/intl';
+import { modifier } from 'ember-modifier';
 
 type Args = {
   controller: SayController;
@@ -16,6 +17,22 @@ export default class NumberInsertComponent extends Component<Args> {
   @tracked label?: string;
   @tracked minimumValue = '';
   @tracked maximumValue = '';
+  @tracked validMinimum = true;
+  @tracked validMaximum = true;
+
+  minimumInput = modifier(
+    (element: HTMLInputElement) => {
+      this.validMinimum = element.checkValidity();
+    },
+    { eager: false },
+  );
+
+  maximumInput = modifier(
+    (element: HTMLInputElement) => {
+      this.validMaximum = element.checkValidity();
+    },
+    { eager: false },
+  );
 
   get controller() {
     return this.args.controller;
@@ -26,6 +43,9 @@ export default class NumberInsertComponent extends Component<Args> {
   }
 
   get numberVariableError() {
+    if (!this.validMinimum || !this.validMaximum) {
+      return this.intl.t('variable.number.error-not-number');
+    }
     const minVal = this.minimumValue;
     const maxVal = this.maximumValue;
     if (
@@ -44,15 +64,17 @@ export default class NumberInsertComponent extends Component<Args> {
     this.label = (event.target as HTMLInputElement).value;
   }
 
-  @action
-  updateMinimumValue(event: InputEvent) {
-    this.minimumValue = (event.target as HTMLInputElement).value;
-  }
+  updateMinimumValue = (event: InputEvent) => {
+    const element = event.target as HTMLInputElement;
+    this.validMinimum = element.checkValidity();
+    this.minimumValue = element.value;
+  };
 
-  @action
-  updateMaximumValue(event: InputEvent) {
-    this.maximumValue = (event.target as HTMLInputElement).value;
-  }
+  updateMaximumValue = (event: InputEvent) => {
+    const element = event.target as HTMLInputElement;
+    this.validMaximum = element.checkValidity();
+    this.maximumValue = element.value;
+  };
 
   @action
   insert() {